Improve return type inference of `applyToPoint`
`applyToPoint` and `applyToPoints` accept points of the form `[1, 2]` or `{ x: 1, y: 2 }`. The shape of the points they return matches the shape that was passed in. However, the previous type definitions did not capture this relationship (that is, "the shape returned matches the shape that was passed in").

I've written more about this at https://github.com/bgschiller/whybother-ts/blob/master/okay-im-convinced/generic-functions/README.md